Post release improvements

+++ apps/backend/src/bot/messages/eventChannelCalendar.message.ts
@@ -35,7 +35,10 @@ export async function buildEventChannelCalendarMessage(id: number, client: Disco
 	const calendarEmbed = new EmbedBuilder()
 		.setColor(PRIMARY_COLOR)
 		.setTitle(":page_with_curl: Upcoming Events")
-		.setDescription("Times are shown in your timezone");
+		.setDescription("Times are shown in your timezone" + (channel.events.length === 0 ? "\n\n*No scheduled events*" : ""))
+		.setFooter({
+			text: "Only events posted in this channel will be listed"
+		});
 	
 	const now = new Date();
 
@@ -126,7 +129,7 @@ async function getEventChannelCalendarMessage(client: DiscordClient, channel: Ev
 	}
 }
 
-export async function updateEventChannelCalendarMessage(client: DiscordClient, channel: EventChannel) {
+export async function updateEventChannelCalendarMessage(client: DiscordClient, channel: EventChannel, repost = false) {
 	try {
 		
 		const payload = await buildEventChannelCalendarMessage(channel.id, client);
@@ -134,18 +137,22 @@ export async function updateEventChannelCalendarMessage(client: DiscordClient, c
 		
 		const discordChannel = await getDiscordChannel(client, channel);
 		let message = await getEventChannelCalendarMessage(client, channel);
-		if (message) {
+		if (repost && message) {
 			await message.delete();
 		}
 		
-		message = await discordChannel.send(payload);
-
-		await database.eventChannel.update({
-			where: { id: channel.id },
-			data: {
-				discordCalendarMessageId: message.id
-			}
-		});
+		if (repost || !message) {
+			message = await discordChannel.send(payload);
+	
+			await database.eventChannel.update({
+				where: { id: channel.id },
+				data: {
+					discordCalendarMessageId: message.id
+				}
+			});
+		} else {
+			await message.edit(payload);
+		}
 	} catch (err) {
 		logger.error("Failed to update event channel calendar message", { channel: channel.id }, err);
 	}
